Saving System
This commint focused on adding a system to save and detect collision

I implemented a json saving system, when an object is deleted or added the save file is changed. When you want to place a new object we check for collisions with all the other objects in the save file. This is way better than the old system because it works 100% of the time. But it needs optimization.

TON of optimization. After just a few objects it gets really slow.