Connecting structures
Now you can click on the green or red cable to create connections between the combinators. The connection will be stored. Is also possible to remove the cables by placing a cable in the same place that already exist and if you delete  a structure all cables and references will be delete with it. The code is very bad but it works. It should not require much more code to be completed.