Cable code rewritte
The cable code was rewritten from scratch.
Now is way more clean and easy to read.
Is should also be easier to change if needed.
This allowed me to also now store where the connection is going other than just the structure, this means that is possible to reconstruct the connections with only the data stored. Even if this still needs to be implemented.